refactor: modularize codebase and add health checks
