+++ Cargo.toml
@@ -10,11 +10,14 @@ clap = { version = "4.5.21", features = ["derive", "env"] }
 dotenvy = "0.15.7"
 futures = "0.3.31"
 hcloud = "0.25.0"
+http-body-util = "0.1"
+hyper = { version = "1.6", features = ["server", "http1"] }
+hyper-util = { version = "0.1", features = ["tokio"] }
 k8s-openapi = { version = "0.27.0", features = ["v1_35"] }
 kube = { version = "3.0.0", features = ["runtime"] }
 kube-leader-election = "0.43.0"
 thiserror = "2.0.3"
-tokio = { version = "1.41.1", features = ["macros", "rt-multi-thread", "signal"] }
+tokio = { version = "1.41.1", features = ["macros", "rt-multi-thread", "signal", "net"] }
 tokio-util = "0.7"
 tower-http = { version = "0.6.6", features = ["util"] }
 tracing = "0.1.40"
