+++ src/lb/mod.rs
@@ -1,252 +1,78 @@
-use async_trait::async_trait;
+//! Load balancer management for Hetzner Cloud.
+//!
+//! This module provides the core load balancer functionality, including:
+//! - Configuration parsing from Kubernetes service annotations
+//! - Reconciliation of load balancer state with Hetzner Cloud
+//! - Cleanup of load balancer resources
+
+mod api;
+mod config;
+mod types;
+
+use std::collections::{HashMap, HashSet};
+
 use hcloud::{
-    apis::{
-        configuration::Configuration as HcloudConfig,
-        load_balancers_api::{
-            AddServiceParams, AddTargetParams, AttachLoadBalancerToNetworkParams,
-            ChangeAlgorithmParams, ChangeTypeOfLoadBalancerParams, DeleteLoadBalancerParams,
-            DeleteServiceParams, DetachLoadBalancerFromNetworkParams, ListLoadBalancersParams,
-            RemoveTargetParams, UpdateServiceParams,
-        },
-        networks_api::ListNetworksParams,
-    },
+    apis::configuration::Configuration as HcloudConfig,
     models::{
-        AttachLoadBalancerToNetworkRequest, ChangeTypeOfLoadBalancerRequest, DeleteServiceRequest,
-        DetachLoadBalancerFromNetworkRequest, LoadBalancerAddTarget, LoadBalancerAlgorithm,
-        LoadBalancerService, LoadBalancerServiceHealthCheck, RemoveTargetRequest,
+        LoadBalancerAlgorithm, LoadBalancerService, LoadBalancerServiceHealthCheck,
         UpdateLoadBalancerService,
     },
 };
 use k8s_openapi::api::core::v1::Service;
-use kube::ResourceExt;
-use std::{
-    collections::{HashMap, HashSet},
-    str::FromStr,
-};
 
 use crate::{
     CurrentContext,
-    config::OperatorConfig,
-    consts,
     error::{RobotLBError, RobotLBResult},
 };
 
-#[derive(Debug)]
-pub struct LBService {
-    pub listen_port: i32,
-    pub target_port: i32,
-}
+pub(crate) use api::{HcloudLoadBalancerApi, LiveHcloudLoadBalancerApi};
+pub(crate) use config::parse_load_balancer_config;
+pub use types::LBService;
+pub(crate) use types::{
+    ServiceReconcileAction, TargetReconcileAction, normalize_ip, service_matches_desired,
+};
 
-enum LBAlgorithm {
-    RoundRobin,
-    LeastConnections,
-}
-
-#[async_trait]
-trait HcloudLoadBalancerApi: Send + Sync {
-    async fn update_service(
-        &self,
-        load_balancer_id: i64,
-        service: UpdateLoadBalancerService,
-    ) -> RobotLBResult<()>;
-
-    async fn delete_service(&self, load_balancer_id: i64, listen_port: i32) -> RobotLBResult<()>;
-
-    async fn add_service(
-        &self,
-        load_balancer_id: i64,
-        service: LoadBalancerService,
-    ) -> RobotLBResult<()>;
-
-    async fn remove_target(&self, load_balancer_id: i64, target_ip: String) -> RobotLBResult<()>;
-
-    async fn add_target(&self, load_balancer_id: i64, target_ip: String) -> RobotLBResult<()>;
-
-    async fn change_algorithm(
-        &self,
-        load_balancer_id: i64,
-        algorithm: LoadBalancerAlgorithm,
-    ) -> RobotLBResult<()>;
-
-    async fn change_type(&self, load_balancer_id: i64, balancer_type: String) -> RobotLBResult<()>;
-}
-
-struct LiveHcloudLoadBalancerApi {
-    hcloud_config: HcloudConfig,
-}
-
-#[async_trait]
-impl HcloudLoadBalancerApi for LiveHcloudLoadBalancerApi {
-    async fn update_service(
-        &self,
-        load_balancer_id: i64,
-        service: UpdateLoadBalancerService,
-    ) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::update_service(
-            &self.hcloud_config,
-            UpdateServiceParams {
-                id: load_balancer_id,
-                body: service,
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn delete_service(&self, load_balancer_id: i64, listen_port: i32) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::delete_service(
-            &self.hcloud_config,
-            DeleteServiceParams {
-                id: load_balancer_id,
-                delete_service_request: DeleteServiceRequest { listen_port },
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn add_service(
-        &self,
-        load_balancer_id: i64,
-        service: LoadBalancerService,
-    ) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::add_service(
-            &self.hcloud_config,
-            AddServiceParams {
-                id: load_balancer_id,
-                body: service,
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn remove_target(&self, load_balancer_id: i64, target_ip: String) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::remove_target(
-            &self.hcloud_config,
-            RemoveTargetParams {
-                id: load_balancer_id,
-                remove_target_request: RemoveTargetRequest {
-                    ip: Some(Box::new(hcloud::models::LoadBalancerTargetIp {
-                        ip: target_ip,
-                    })),
-                    ..Default::default()
-                },
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn add_target(&self, load_balancer_id: i64, target_ip: String) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::add_target(
-            &self.hcloud_config,
-            AddTargetParams {
-                id: load_balancer_id,
-                body: LoadBalancerAddTarget {
-                    ip: Some(Box::new(hcloud::models::LoadBalancerTargetIp {
-                        ip: target_ip,
-                    })),
-                    ..Default::default()
-                },
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn change_algorithm(
-        &self,
-        load_balancer_id: i64,
-        algorithm: LoadBalancerAlgorithm,
-    ) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::change_algorithm(
-            &self.hcloud_config,
-            ChangeAlgorithmParams {
-                id: load_balancer_id,
-                body: algorithm,
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn change_type(&self, load_balancer_id: i64, balancer_type: String) -> RobotLBResult<()> {
-        hcloud::apis::load_balancers_api::change_type_of_load_balancer(
-            &self.hcloud_config,
-            ChangeTypeOfLoadBalancerParams {
-                id: load_balancer_id,
-                change_type_of_load_balancer_request: ChangeTypeOfLoadBalancerRequest {
-                    load_balancer_type: balancer_type,
-                },
-            },
-        )
-        .await?;
-        Ok(())
-    }
-}
-
-/// Struct representing a load balancer
+/// Struct representing a load balancer.
+///
 /// It holds all the necessary information to manage the load balancer
 /// in Hetzner Cloud.
 #[derive(Debug)]
 pub struct LoadBalancer {
+    /// Name of the load balancer.
     pub name: String,
+    /// Services exposed by the load balancer (`listen_port` -> `target_port`).
     pub services: HashMap<i32, i32>,
+    /// Target IPs for the load balancer.
     pub targets: Vec<String>,
+    /// Optional private IP for the load balancer.
     pub private_ip: Option<String>,
 
+    /// Health check interval in seconds.
     pub check_interval: i32,
+    /// Health check timeout in seconds.
     pub timeout: i32,
+    /// Number of health check retries.
     pub retries: i32,
+    /// Whether proxy protocol is enabled.
     pub proxy_mode: bool,
 
+    /// Hetzner Cloud location.
     pub location: String,
+    /// Load balancer type (e.g., "lb11").
     pub balancer_type: String,
+    /// Load balancing algorithm.
     pub algorithm: LoadBalancerAlgorithm,
+    /// Optional network name for private networking.
     pub network_name: Option<String>,
 
+    /// Hetzner Cloud API configuration.
     pub hcloud_config: HcloudConfig,
-}
-
-#[derive(Debug)]
-struct ParsedLoadBalancerConfig {
-    name: String,
-    private_ip: Option<String>,
-    balancer_type: String,
-    check_interval: i32,
-    timeout: i32,
-    retries: i32,
-    location: String,
-    proxy_mode: bool,
-    network_name: Option<String>,
-    algorithm: LoadBalancerAlgorithm,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ServiceReconcileAction {
-    Update {
-        listen_port: i32,
-        destination_port: i32,
-    },
-    Delete {
-        listen_port: i32,
-    },
-    Add {
-        listen_port: i32,
-        destination_port: i32,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum TargetReconcileAction {
-    Remove { target_ip: String },
-    Add { target_ip: String },
 }
 
 impl LoadBalancer {
     /// Create a new `LoadBalancer` instance from a Kubernetes service
     /// and the current context.
+    ///
     /// This method will try to extract all the necessary information
     /// from the service annotations and the context.
     /// If some of the required information is missing, the method will
@@ -276,6 +102,7 @@ impl LoadBalancer {
     }
 
     /// Add a service to the load balancer.
+    ///
     /// The service will listen on the `listen_port` and forward the
     /// traffic to the `target_port` to all targets.
     pub fn add_service(&mut self, listen_port: i32, target_port: i32) {
@@ -283,6 +110,7 @@ impl LoadBalancer {
     }
 
     /// Add a target to the load balancer.
+    ///
     /// The target will receive the traffic from the services.
     /// The target is identified by its IP address.
     pub fn add_target(&mut self, ip: &str) {
@@ -294,7 +122,7 @@ impl LoadBalancer {
     /// # Errors
     ///
     /// Returns an error if the Hetzner Cloud API call fails.
-    #[tracing::instrument(skip(self), fields(lb_name=self.name))]
+    #[tracing::instrument(skip(self), fields(lb_name = self.name))]
     pub async fn reconcile(&self) -> RobotLBResult<hcloud::models::LoadBalancer> {
         let api = LiveHcloudLoadBalancerApi {
             hcloud_config: self.hcloud_config.clone(),
@@ -309,6 +137,7 @@ impl LoadBalancer {
     }
 
     /// Reconcile the services of the load balancer.
+    ///
     /// This method will compare the desired configuration of the services
     /// with the current configuration of the services in the load balancer.
     /// If the configuration does not match, the method will update the service.
@@ -331,7 +160,9 @@ impl LoadBalancer {
                         hcloud_balancer.id,
                         UpdateLoadBalancerService {
                             http: None,
-                            protocol: Some(hcloud::models::update_load_balancer_service::Protocol::Tcp),
+                            protocol: Some(
+                                hcloud::models::update_load_balancer_service::Protocol::Tcp,
+                            ),
                             listen_port,
                             destination_port: Some(destination_port),
                             proxyprotocol: Some(self.proxy_mode),
@@ -394,6 +225,7 @@ impl LoadBalancer {
     }
 
     /// Reconcile the targets of the load balancer.
+    ///
     /// This method will compare the desired configuration of the targets
     /// with the current configuration of the targets in the load balancer.
     /// If the configuration does not match, the method will update the target.
@@ -418,6 +250,7 @@ impl LoadBalancer {
     }
 
     /// Reconcile the load balancer algorithm.
+    ///
     /// This method will compare the desired algorithm configuration
     /// and update it if it does not match the current configuration.
     async fn reconcile_algorithm(
@@ -461,7 +294,14 @@ impl LoadBalancer {
 
         for service in &hcloud_balancer.services {
             if let Some(destination_port) = self.services.get(&service.listen_port) {
-                if !self.service_matches_desired(service, *destination_port) {
+                if !service_matches_desired(
+                    service,
+                    *destination_port,
+                    self.check_interval,
+                    self.retries,
+                    self.timeout,
+                    self.proxy_mode,
+                ) {
                     actions.push(ServiceReconcileAction::Update {
                         listen_port: service.listen_port,
                         destination_port: *destination_port,
@@ -498,22 +338,6 @@ impl LoadBalancer {
 
         actions.extend(missing_services);
         actions
-    }
-
-    fn service_matches_desired(
-        &self,
-        service: &LoadBalancerService,
-        destination_port: i32,
-    ) -> bool {
-        service.destination_port == destination_port
-            && service.health_check.port == destination_port
-            && service.health_check.interval == self.check_interval
-            && service.health_check.retries == self.retries
-            && service.health_check.timeout == self.timeout
-            && service.proxyprotocol == self.proxy_mode
-            && service.http.is_none()
-            && service.health_check.protocol
-                == hcloud::models::load_balancer_service_health_check::Protocol::Tcp
     }
 
     fn plan_target_actions(
@@ -579,6 +403,7 @@ impl LoadBalancer {
     }
 
     /// Reconcile the network of the load balancer.
+    ///
     /// This method will compare the desired network configuration
     /// with the current network configuration of the load balancer.
     /// If the configuration does not match, the method will update the
@@ -587,7 +412,7 @@ impl LoadBalancer {
         &self,
         hcloud_balancer: &hcloud::models::LoadBalancer,
     ) -> RobotLBResult<()> {
-        // If the network name is not provided, and laod balancer is not attached to any network,
+        // If the network name is not provided, and load balancer is not attached to any network,
         // we can skip this step.
         if self.network_name.is_none() && hcloud_balancer.private_net.is_empty() {
             return Ok(());
@@ -618,17 +443,8 @@ impl LoadBalancer {
                     }
                 }
                 tracing::info!("Detaching balancer from network {}", private_net_id);
-                hcloud::apis::load_balancers_api::detach_load_balancer_from_network(
-                    &self.hcloud_config,
-                    DetachLoadBalancerFromNetworkParams {
-                        id: hcloud_balancer.id,
-                        detach_load_balancer_from_network_request:
-                            DetachLoadBalancerFromNetworkRequest {
-                                network: private_net_id,
-                            },
-                    },
-                )
-                .await?;
+                api::detach_from_network(&self.hcloud_config, hcloud_balancer.id, private_net_id)
+                    .await?;
             }
         }
         if !contain_desired_network {
@@ -636,16 +452,11 @@ impl LoadBalancer {
                 return Ok(());
             };
             tracing::info!("Attaching balancer to network {}", network_id);
-            hcloud::apis::load_balancers_api::attach_load_balancer_to_network(
+            api::attach_to_network(
                 &self.hcloud_config,
-                AttachLoadBalancerToNetworkParams {
-                    id: hcloud_balancer.id,
-                    attach_load_balancer_to_network_request: AttachLoadBalancerToNetworkRequest {
-                        ip: self.private_ip.clone(),
-                        ip_range: None,
-                        network: network_id,
-                    },
-                },
+                hcloud_balancer.id,
+                network_id,
+                self.private_ip.clone(),
             )
             .await?;
         }
@@ -653,6 +464,7 @@ impl LoadBalancer {
     }
 
     /// Cleanup the load balancer.
+    ///
     /// This method will remove all the services and targets from the
     /// load balancer.
     ///
@@ -660,7 +472,8 @@ impl LoadBalancer {
     ///
     /// Returns an error if the Hetzner Cloud API call fails.
     pub async fn cleanup(&self) -> RobotLBResult<()> {
-        let Some(hcloud_balancer) = self.get_hcloud_lb().await? else {
+        let Some(hcloud_balancer) = api::get_load_balancer(&self.hcloud_config, &self.name).await?
+        else {
             return Ok(());
         };
         for service in &hcloud_balancer.services {
@@ -669,73 +482,33 @@ impl LoadBalancer {
                 service.listen_port,
                 hcloud_balancer.name,
             );
-            hcloud::apis::load_balancers_api::delete_service(
-                &self.hcloud_config,
-                DeleteServiceParams {
-                    id: hcloud_balancer.id,
-                    delete_service_request: DeleteServiceRequest {
-                        listen_port: service.listen_port,
-                    },
-                },
-            )
-            .await?;
+            api::delete_service(&self.hcloud_config, hcloud_balancer.id, service.listen_port)
+                .await?;
         }
         for target in &hcloud_balancer.targets {
             if let Some(target_ip) = target.ip.clone() {
                 tracing::info!("Removing target {}", target_ip.ip);
-                hcloud::apis::load_balancers_api::remove_target(
-                    &self.hcloud_config,
-                    RemoveTargetParams {
-                        id: hcloud_balancer.id,
-                        remove_target_request: RemoveTargetRequest {
-                            ip: Some(target_ip),
-                            ..Default::default()
-                        },
-                    },
-                )
-                .await?;
+                api::remove_target(&self.hcloud_config, hcloud_balancer.id, *target_ip).await?;
             }
         }
-        hcloud::apis::load_balancers_api::delete_load_balancer(
-            &self.hcloud_config,
-            DeleteLoadBalancerParams {
-                id: hcloud_balancer.id,
-            },
-        )
-        .await?;
+        api::delete_load_balancer(&self.hcloud_config, hcloud_balancer.id).await?;
         Ok(())
     }
 
     /// Get the load balancer from Hetzner Cloud.
+    ///
     /// This method will try to find the load balancer with the name
     /// specified in the `LoadBalancer` struct.
     ///
     /// The method might return an error if the load balancer is not found
     /// or if there are multiple load balancers with the same name.
     async fn get_hcloud_lb(&self) -> RobotLBResult<Option<hcloud::models::LoadBalancer>> {
-        let hcloud_balancers = hcloud::apis::load_balancers_api::list_load_balancers(
-            &self.hcloud_config,
-            ListLoadBalancersParams {
-                name: Some(self.name.clone()),
-                ..Default::default()
-            },
-        )
-        .await?;
-        if hcloud_balancers.load_balancers.len() > 1 {
-            tracing::warn!(
-                "Found more than one balancer with name {}, skipping",
-                self.name
-            );
-            return Err(RobotLBError::SkipService);
-        }
-        // Here we just return the first load balancer,
-        // if it exists, otherwise we return None
-        Ok(hcloud_balancers.load_balancers.into_iter().next())
+        api::get_load_balancer(&self.hcloud_config, &self.name).await
     }
 
     /// Get or create the load balancer in Hetzner Cloud.
     ///
-    /// this method will try to find the load balancer with the name
+    /// This method will try to find the load balancer with the name
     /// specified in the `LoadBalancer` struct. If the load balancer
     /// is not found, the method will create a new load balancer
     /// with the specified configuration in service's annotations.
@@ -745,35 +518,18 @@ impl LoadBalancer {
             return Ok(balancer);
         }
 
-        let response = hcloud::apis::load_balancers_api::create_load_balancer(
+        api::create_load_balancer(
             &self.hcloud_config,
-            hcloud::apis::load_balancers_api::CreateLoadBalancerParams {
-                create_load_balancer_request: hcloud::models::CreateLoadBalancerRequest {
-                    algorithm: Some(Box::new(self.algorithm.clone())),
-                    labels: None,
-                    load_balancer_type: self.balancer_type.clone(),
-                    location: Some(self.location.clone()),
-                    name: self.name.clone(),
-                    network: None,
-                    network_zone: None,
-                    public_interface: Some(true),
-                    services: Some(vec![]),
-                    targets: Some(vec![]),
-                },
-            },
+            &self.name,
+            &self.location,
+            &self.balancer_type,
+            self.algorithm.clone(),
         )
-        .await;
-        if let Err(e) = response {
-            tracing::error!("Failed to create load balancer: {:?}", e);
-            return Err(RobotLBError::HCloudError(format!(
-                "Failed to create load balancer: {e:?}"
-            )));
-        }
-
-        Ok(*response.unwrap().load_balancer)
+        .await
     }
 
     /// Get the network from Hetzner Cloud.
+    ///
     /// This method will try to find the network with the name
     /// specified in the `LoadBalancer` struct. It returns `None` only
     /// in case the network name is not provided. If the network was not found,
@@ -782,14 +538,7 @@ impl LoadBalancer {
         let Some(network_name) = self.network_name.clone() else {
             return Ok(None);
         };
-        let response = hcloud::apis::networks_api::list_networks(
-            &self.hcloud_config,
-            ListNetworksParams {
-                name: Some(network_name.clone()),
-                ..Default::default()
-            },
-        )
-        .await?;
+        let response = api::list_networks(&self.hcloud_config, Some(network_name.clone())).await?;
 
         if response.networks.len() > 1 {
             tracing::warn!(
@@ -811,142 +560,12 @@ impl LoadBalancer {
     }
 }
 
-fn parse_load_balancer_config(
-    svc: &Service,
-    config: &OperatorConfig,
-) -> RobotLBResult<ParsedLoadBalancerConfig> {
-    let retries = svc
-        .annotations()
-        .get(consts::LB_RETRIES_ANN_NAME)
-        .map(String::as_str)
-        .map(i32::from_str)
-        .transpose()?
-        .unwrap_or(config.default_lb_retries);
-
-    let timeout = svc
-        .annotations()
-        .get(consts::LB_TIMEOUT_ANN_NAME)
-        .map(String::as_str)
-        .map(i32::from_str)
-        .transpose()?
-        .unwrap_or(config.default_lb_timeout);
-
-    let check_interval = svc
-        .annotations()
-        .get(consts::LB_CHECK_INTERVAL_ANN_NAME)
-        .map(String::as_str)
-        .map(i32::from_str)
-        .transpose()?
-        .unwrap_or(config.default_lb_interval);
-
-    let proxy_mode = svc
-        .annotations()
-        .get(consts::LB_PROXY_MODE_LABEL_NAME)
-        .map(String::as_str)
-        .map(bool::from_str)
-        .transpose()?
-        .unwrap_or(config.default_lb_proxy_mode_enabled);
-
-    let location = svc
-        .annotations()
-        .get(consts::LB_LOCATION_LABEL_NAME)
-        .cloned()
-        .unwrap_or_else(|| config.default_lb_location.clone());
-
-    let balancer_type = svc
-        .annotations()
-        .get(consts::LB_BALANCER_TYPE_LABEL_NAME)
-        .cloned()
-        .unwrap_or_else(|| config.default_balancer_type.clone());
-
-    let algorithm = svc
-        .annotations()
-        .get(consts::LB_ALGORITHM_LABEL_NAME)
-        .map(String::as_str)
-        .or(Some(&config.default_lb_algorithm))
-        .map(LBAlgorithm::from_str)
-        .transpose()?
-        .unwrap_or(LBAlgorithm::LeastConnections);
-
-    let network_name = svc
-        .annotations()
-        .get(consts::LB_NETWORK_LABEL_NAME)
-        .or(config.default_network.as_ref())
-        .cloned();
-
-    let name = svc
-        .annotations()
-        .get(consts::LB_NAME_LABEL_NAME)
-        .cloned()
-        .unwrap_or_else(|| svc.name_any());
-
-    let private_ip = svc
-        .annotations()
-        .get(consts::LB_PRIVATE_IP_LABEL_NAME)
-        .cloned();
-
-    Ok(ParsedLoadBalancerConfig {
-        name,
-        private_ip,
-        balancer_type,
-        check_interval,
-        timeout,
-        retries,
-        location,
-        proxy_mode,
-        network_name,
-        algorithm: algorithm.into(),
-    })
-}
-
-impl FromStr for LBAlgorithm {
-    type Err = RobotLBError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "round-robin" => Ok(Self::RoundRobin),
-            "least-connections" => Ok(Self::LeastConnections),
-            _ => Err(RobotLBError::UnknownLBAlgorithm),
-        }
-    }
-}
-
-impl From<LBAlgorithm> for LoadBalancerAlgorithm {
-    fn from(value: LBAlgorithm) -> Self {
-        let r#type = match value {
-            LBAlgorithm::RoundRobin => hcloud::models::load_balancer_algorithm::Type::RoundRobin,
-            LBAlgorithm::LeastConnections => {
-                hcloud::models::load_balancer_algorithm::Type::LeastConnections
-            }
-        };
-        Self { r#type }
-    }
-}
-
-fn normalize_ip(ip: &str) -> String {
-    ip.split('/').next().unwrap_or(ip).to_string()
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{
-        HcloudLoadBalancerApi, LoadBalancer, ServiceReconcileAction, TargetReconcileAction,
-        parse_load_balancer_config,
-    };
-    use crate::{config::OperatorConfig, consts};
+    use super::*;
     use async_trait::async_trait;
-    use hcloud::{
-        apis::configuration::Configuration as HcloudConfig,
-        models::{
-            LoadBalancerAlgorithm, LoadBalancerService, LoadBalancerServiceHealthCheck,
-            UpdateLoadBalancerService,
-        },
-    };
-    use k8s_openapi::{api::core::v1::Service, apimachinery::pkg::apis::meta::v1::ObjectMeta};
-    use std::{
-        collections::{BTreeMap, HashMap},
-        sync::Mutex,
-    };
-    use tracing::level_filters::LevelFilter;
+    use hcloud::apis::configuration::Configuration as HcloudConfig;
+    use std::sync::Mutex;
 
     #[derive(Debug, Clone, PartialEq, Eq)]
     enum ApiCall {
@@ -993,7 +612,7 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             service: UpdateLoadBalancerService,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1010,7 +629,7 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             listen_port: i32,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1022,7 +641,7 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             service: LoadBalancerService,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1037,7 +656,7 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             target_ip: String,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1045,11 +664,7 @@ mod tests {
             Ok(())
         }
 
-        async fn add_target(
-            &self,
-            _load_balancer_id: i64,
-            target_ip: String,
-        ) -> crate::error::RobotLBResult<()> {
+        async fn add_target(&self, _load_balancer_id: i64, target_ip: String) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1061,7 +676,7 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             _algorithm: LoadBalancerAlgorithm,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
@@ -1073,51 +688,12 @@ mod tests {
             &self,
             _load_balancer_id: i64,
             balancer_type: String,
-        ) -> crate::error::RobotLBResult<()> {
+        ) -> RobotLBResult<()> {
             self.calls
                 .lock()
                 .expect("lock should not be poisoned")
                 .push(ApiCall::ChangeType { balancer_type });
             Ok(())
-        }
-    }
-
-    fn base_config() -> OperatorConfig {
-        OperatorConfig {
-            hcloud_token: "token".to_string(),
-            default_network: Some("default-net".to_string()),
-            dynamic_node_selector: true,
-            default_lb_retries: 3,
-            default_lb_timeout: 10,
-            default_lb_interval: 15,
-            default_lb_location: "hel1".to_string(),
-            default_balancer_type: "lb11".to_string(),
-            default_lb_algorithm: "least-connections".to_string(),
-            default_lb_proxy_mode_enabled: false,
-            ipv6_ingress: false,
-            leader_election_namespace: None,
-            leader_election_lease_name: "robotlb-leader-election".to_string(),
-            leader_election_lease_ttl_secs: 15,
-            leader_election_renew_interval_secs: 5,
-            log_level: LevelFilter::INFO,
-        }
-    }
-
-    fn service_with_annotations(
-        annotations: impl IntoIterator<Item = (&'static str, &'static str)>,
-    ) -> Service {
-        let annotation_map = annotations
-            .into_iter()
-            .map(|(k, v)| (k.to_string(), v.to_string()))
-            .collect::<BTreeMap<_, _>>();
-
-        Service {
-            metadata: ObjectMeta {
-                name: Some("svc-name".to_string()),
-                annotations: Some(annotation_map),
-                ..Default::default()
-            },
-            ..Default::default()
         }
     }
 
@@ -1186,69 +762,6 @@ mod tests {
             }),
             ..Default::default()
         }
-    }
-
-    #[test]
-    fn uses_defaults_when_annotations_are_missing() {
-        let config = base_config();
-        let svc = service_with_annotations([]);
-
-        let parsed = parse_load_balancer_config(&svc, &config).expect("parse should succeed");
-
-        assert_eq!(parsed.name, "svc-name");
-        assert_eq!(parsed.retries, 3);
-        assert_eq!(parsed.timeout, 10);
-        assert_eq!(parsed.check_interval, 15);
-        assert_eq!(parsed.location, "hel1");
-        assert_eq!(parsed.balancer_type, "lb11");
-        assert_eq!(parsed.network_name.as_deref(), Some("default-net"));
-        assert!(!parsed.proxy_mode);
-    }
-
-    #[test]
-    fn parses_annotations_into_load_balancer_config() {
-        let mut config = base_config();
-        config.default_network = None;
-        let svc = service_with_annotations([
-            (consts::LB_NAME_LABEL_NAME, "custom-lb"),
-            (consts::LB_RETRIES_ANN_NAME, "5"),
-            (consts::LB_TIMEOUT_ANN_NAME, "8"),
-            (consts::LB_CHECK_INTERVAL_ANN_NAME, "20"),
-            (consts::LB_PROXY_MODE_LABEL_NAME, "true"),
-            (consts::LB_LOCATION_LABEL_NAME, "nbg1"),
-            (consts::LB_BALANCER_TYPE_LABEL_NAME, "lb31"),
-            (consts::LB_ALGORITHM_LABEL_NAME, "round-robin"),
-            (consts::LB_NETWORK_LABEL_NAME, "private-net"),
-            (consts::LB_PRIVATE_IP_LABEL_NAME, "10.10.0.5"),
-        ]);
-
-        let parsed = parse_load_balancer_config(&svc, &config).expect("parse should succeed");
-
-        assert_eq!(parsed.name, "custom-lb");
-        assert_eq!(parsed.retries, 5);
-        assert_eq!(parsed.timeout, 8);
-        assert_eq!(parsed.check_interval, 20);
-        assert_eq!(parsed.location, "nbg1");
-        assert_eq!(parsed.balancer_type, "lb31");
-        assert_eq!(parsed.network_name.as_deref(), Some("private-net"));
-        assert_eq!(parsed.private_ip.as_deref(), Some("10.10.0.5"));
-        assert!(parsed.proxy_mode);
-        assert_eq!(
-            parsed.algorithm.r#type,
-            hcloud::models::load_balancer_algorithm::Type::RoundRobin
-        );
-    }
-
-    #[test]
-    fn returns_error_for_invalid_algorithm_annotation() {
-        let config = base_config();
-        let svc = service_with_annotations([(consts::LB_ALGORITHM_LABEL_NAME, "weighted")]);
-
-        let result = parse_load_balancer_config(&svc, &config);
-        assert!(matches!(
-            result,
-            Err(crate::error::RobotLBError::UnknownLBAlgorithm)
-        ));
     }
 
     #[tokio::test]
